refactor(dat): extract Production mode to DatProductionComposer (#611 PR2)
Move ComposeProduction, ProductionRowValues, and Production header
building from DatComposer to DatProductionComposer. Add shared
helpers (ApplyConvention, EffectiveNow, MakeRecord,
GetFamilyIdentifiers) to DatComposerShared. DatComposer now
delegates ProductionSet mode to DatProductionComposer. DatComposer.cs
653 to 501 lines. No behavior change.
